20220907问题修复

+++ app/src/main/java/com/pro/maluli/common/entity/JoinLiveEntity.java
@@ -258,7 +258,7 @@ public class JoinLiveEntity implements Serializable {
             private String name;
             private boolean ionotify;
             private boolean muted;
-            private int roomid;
+            private long roomid;
             private String announcement;
             private int queuelevel;
 
@@ -318,11 +318,11 @@ public class JoinLiveEntity implements Serializable {
                 this.muted = muted;
             }
 
-            public int getRoomid() {
+            public long getRoomid() {
                 return roomid;
             }
 
-            public void setRoomid(int roomid) {
+            public void setRoomid(long roomid) {
                 this.roomid = roomid;
             }
